refactor: move v0.1 rpc tests into module
